--- Source/Model/EKTransformAnimation.swift
@@ -1,16 +1,0 @@
-//
-//  EKTransformAnimation.swift
-//  SwiftEntryKit
-//
-//  Created by Daniel Huri on 5/31/18.
-//  Copyright © 2018 CocoaPods. All rights reserved.
-//
-
-import Foundation
-
-struct EKTransformAnimation {
-    
-    let duration: TimeInterval
-    let delay: TimeInterval
-    
-}
